refactor(config): share group policy evaluation
Preserve flat candidate selection and fallback rules through the scope owner. Focused group-policy suites: 96 tests passed; assertion-baseline shrink remains a campaign gate blocker.
