+++ src/config/group-scope-tree.ts
@@ -1,7 +1,8 @@
 // Resolves canonical group policy scopes prepared by channel plugins.
 import { normalizeLowercaseStringOrEmpty } from "@openclaw/normalization-core/string-coerce";
 import type { ChannelId } from "../channels/plugins/channel-id.types.js";
-import { resolveChannelGroups, resolveToolsBySender } from "./group-policy.js";
+import { resolveChannelGroups } from "./channel-groups.js";
+import { resolveToolsBySender } from "./tools-by-sender.js";
 import type { OpenClawConfig } from "./types.openclaw.js";
 import type { GroupToolPolicyBySenderConfig, GroupToolPolicyConfig } from "./types.tools.js";
